requirements.txt
streamlit
pandas
psycopg2-binary
python-dotenv
google-generativeai
langchain
langchain_core
langchain_community